+++ core/src/main/java/cz/zcu/kiv/nlp/ir/Main.java
@@ -21,7 +21,7 @@ import cz.zcu.kiv.nlp.ir.fileLoader.UrlFileLoader;
 import cz.zcu.kiv.nlp.ir.index.Index;
 import cz.zcu.kiv.nlp.ir.index.Indexable;
 import cz.zcu.kiv.nlp.ir.index.TfIdfIndex;
-import cz.zcu.kiv.nlp.ir.index.query.QueryParser;
+import cz.zcu.kiv.nlp.ir.index.query.DefaultQueryParser;
 import cz.zcu.kiv.nlp.ir.index.query.SearchModel;
 import cz.zcu.kiv.nlp.ir.preprocess.DefaultPreprocessor;
 import cz.zcu.kiv.nlp.ir.preprocess.Preprocessor;
@@ -43,9 +43,6 @@ public class Main {
   private static final long DEFAULT_CRAWLER_POLITENESS_INTERVAL = 1200;
 
   public static void main(final String[] args) {
-    QueryParser queryParser = new QueryParser();
-    final var parsed = queryParser.parse("Praha");
-
     final var parser = new CommandParser(LoggerFactory.getILoggerFactory());
     final var config = parser.parse(args);
 
@@ -59,8 +56,9 @@ public class Main {
       return;
     }
 
+    // TODO use custom preprocessor in query parser??
     final Preprocessor preprocessor = createPreprocessor();
-    final Index index = new TfIdfIndex(preprocessor);
+    final Index index = new TfIdfIndex(preprocessor, new DefaultQueryParser());
     if (!index.hasData()) {
       final Storage<? extends Article> storage = config.getStorage();
       if (!storage.hasData()) {
@@ -146,7 +144,8 @@ public class Main {
     final var model = input.getOptionValue()
         .map((option) -> SearchModel.valueOf(option.toUpperCase()))
         .orElseThrow(() -> new IllegalStateException("No query model provided"));
-    final var result = index.search(query, model); // TODO model as parameter
+
+    final var result = index.search(query, model);
     printResult(result, index);
   }
